Added RecommendationsActivity.java: This file contains the main activity responsible for displaying recommendations based on the user's location. It initializes the RecyclerView for displaying recommendations, retrieves the current latitude and longitude, and fetches recommendations using the Places API. It makes use of the fetchPlacesAndDisplay() method to fetch and display places of interest and places to eat separately, and updates the RecyclerView with the fetched places.
RecommendationAdapter.java:
This file is responsible for binding the data of places to the RecyclerView. It defines two constants for header and item view types and contains a list of recommendations that include both headers (Strings) and Place items (Place objects). The adapter inflates the appropriate layout for header and item view types and binds the data to the ViewHolder based on the item's position.

PlacesApiHelper.java:
This helper class contains methods for interacting with the Google Places API. It includes methods for fetching nearby places, fetching place details, and providing callback interfaces to handle the responses from the API. This class abstracts the API calls and makes it easier to interact with the API from other parts of the app.

Place.java:
This file contains the Place class, which is a simple data model representing a place with properties like name, latitude, longitude, distance, and placeId. It also provides a method to calculate the distance between two places based on their latitude and longitude values.

Added recommendation_item.xml file and activity_recommendations.xml and header_item.xml and recommendation_item.xml for design of app